app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
flask-migrate==4.0.5
flask-cors==4.0.0
prometheus-flask-exporter==0.22.0
